app.js actualizado COMPONENTE PRODUCT BORRADO ProductDetail.jsx FUNCIONANDO DE NUEVO
