+++ frontend/src/App.js
@@ -5,7 +5,7 @@ import Landing from "./components/Landing/Landing";
 import NavBar from "./components/NavBar/NavBar";
 import Home from "./views/Home/Home";
 import Footer from "./components/Footer/Footer";
-import Product from "./components/Product/Product";
+// import Product from "./components/Product/Product";
 import Ofertas from "./views/Ofertas/Ofertas";
 
 const App = () => {
@@ -14,10 +14,10 @@ const App = () => {
       <NavBar /> 
       <Switch>
         <Route path="/home" component={Home} />
-        <Route path="/product" component={Product} />
+        {/* <Route path="/product" component={Product} /> COMPONENTE PRODUCT ELIMINADO*/} 
         <Route path="/products/ofertas" component={Ofertas} />
         <Route path="/product/:id" component={ProductDetail} />
-        <Route exact path="/" component={Landing} /> {/* Usamos "exact" aquí */}
+        <Route exact path="/" component={Landing} /> 
       </Switch>
       <Footer />
     </Router>
