+++ frontend/src/views/ProductDetail.jsx
@@ -18,6 +18,14 @@ const ProductDetail = () => {
         return <div>Cargando...</div>;
     }
 
+    const handleAddToCart = () => {
+        // Agregar el producto al carrito
+      };
+    
+    const handleBuyNow = () => {
+        // Comprar el producto ahora
+      };
+
     return (
         <div>
             <h2>Detalles del producto</h2>
@@ -31,6 +39,23 @@ const ProductDetail = () => {
             <p>Categoría: {product.category}</p>
             <p>Subcategoría: {product.subcategory}</p>
             <p>Descripción: {product.description}</p>
+
+            <button onClick={handleAddToCart}>Agregar al carrito</button>
+            <button onClick={handleBuyNow}>Comprar ahora</button>
+
+            {/* <form>
+                <label htmlFor="comment">Comentario:</label>
+                <textarea id="comment" name="comment" />
+                <button type="submit">Enviar comentario</button>
+            </form>
+            <div className="rating">
+                <span className="star">&#9733;</span>
+                <span className="star">&#9733;</span>
+                <span className="star">&#9733;</span>
+                <span className="star">&#9734;</span>
+                <span className="star">&#9734;</span>
+                <p>Calificación promedio: 3 estrellas</p>
+            </div> */}
         </div>
     );
 };
